fix: move class to `AccessControl` folder
